+++ packages/widgetbook/lib/src/rendering/rendering_state.dart
@@ -1,6 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:widgetbook/src/rendering/builders/text_scale_builder.dart';
-
 import 'package:widgetbook/src/rendering/rendering.dart';
 import 'package:widgetbook_models/widgetbook_models.dart';
 
@@ -14,6 +13,7 @@ class RenderingState<CustomTheme> with _$RenderingState<CustomTheme> {
     required LocalizationBuilderFunction localizationBuilder,
     required ThemeBuilderFunction<CustomTheme> themeBuilder,
     required ScaffoldBuilderFunction scaffoldBuilder,
+    required AppBuilderFunction appBuilder,
     required TextScaleBuilder textScaleBuilder,
     required UseCaseBuilderFunction useCaseBuilder,
   }) = _RenderingState;