+++ packages/widgetbook_generator/lib/models/widgetbook_device_frame_data.freezed.dart
@@ -1,5 +1,6 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'widgetbook_device_frame_data.dart';
@@ -11,35 +12,12 @@ part of 'widgetbook_device_frame_data.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
 WidgetbookDeviceFrameData _$WidgetbookDeviceFrameDataFromJson(
     Map<String, dynamic> json) {
   return _WidgetbookDeviceFrameData.fromJson(json);
 }
-
-/// @nodoc
-class _$WidgetbookDeviceFrameDataTearOff {
-  const _$WidgetbookDeviceFrameDataTearOff();
-
-  _WidgetbookDeviceFrameData call(
-      {required String name,
-      required String importStatement,
-      required List<String> dependencies}) {
-    return _WidgetbookDeviceFrameData(
-      name: name,
-      importStatement: importStatement,
-      dependencies: dependencies,
-    );
-  }
-
-  WidgetbookDeviceFrameData fromJson(Map<String, Object?> json) {
-    return WidgetbookDeviceFrameData.fromJson(json);
-  }
-}
-
-/// @nodoc
-const $WidgetbookDeviceFrameData = _$WidgetbookDeviceFrameDataTearOff();
 
 /// @nodoc
 mixin _$WidgetbookDeviceFrameData {
@@ -94,26 +72,28 @@ class _$WidgetbookDeviceFrameDataCopyWithImpl<$Res>
 }
 
 /// @nodoc
-abstract class _$WidgetbookDeviceFrameDataCopyWith<$Res>
+abstract class _$$_WidgetbookDeviceFrameDataCopyWith<$Res>
     implements $WidgetbookDeviceFrameDataCopyWith<$Res> {
-  factory _$WidgetbookDeviceFrameDataCopyWith(_WidgetbookDeviceFrameData value,
-          $Res Function(_WidgetbookDeviceFrameData) then) =
-      __$WidgetbookDeviceFrameDataCopyWithImpl<$Res>;
+  factory _$$_WidgetbookDeviceFrameDataCopyWith(
+          _$_WidgetbookDeviceFrameData value,
+          $Res Function(_$_WidgetbookDeviceFrameData) then) =
+      __$$_WidgetbookDeviceFrameDataCopyWithImpl<$Res>;
   @override
   $Res call({String name, String importStatement, List<String> dependencies});
 }
 
 /// @nodoc
-class __$WidgetbookDeviceFrameDataCopyWithImpl<$Res>
+class __$$_WidgetbookDeviceFrameDataCopyWithImpl<$Res>
     extends _$WidgetbookDeviceFrameDataCopyWithImpl<$Res>
-    implements _$WidgetbookDeviceFrameDataCopyWith<$Res> {
-  __$WidgetbookDeviceFrameDataCopyWithImpl(_WidgetbookDeviceFrameData _value,
-      $Res Function(_WidgetbookDeviceFrameData) _then)
-      : super(_value, (v) => _then(v as _WidgetbookDeviceFrameData));
+    implements _$$_WidgetbookDeviceFrameDataCopyWith<$Res> {
+  __$$_WidgetbookDeviceFrameDataCopyWithImpl(
+      _$_WidgetbookDeviceFrameData _value,
+      $Res Function(_$_WidgetbookDeviceFrameData) _then)
+      : super(_value, (v) => _then(v as _$_WidgetbookDeviceFrameData));
 
   @override
-  _WidgetbookDeviceFrameData get _value =>
-      super._value as _WidgetbookDeviceFrameData;
+  _$_WidgetbookDeviceFrameData get _value =>
+      super._value as _$_WidgetbookDeviceFrameData;
 
   @override
   $Res call({
@@ -121,7 +101,7 @@ class __$WidgetbookDeviceFrameDataCopyWithImpl<$Res>
     Object? importStatement = freezed,
     Object? dependencies = freezed,
   }) {
-    return _then(_WidgetbookDeviceFrameData(
+    return _then(_$_WidgetbookDeviceFrameData(
       name: name == freezed
           ? _value.name
           : name // ignore: cast_nullable_to_non_nullable
@@ -131,7 +111,7 @@ class __$WidgetbookDeviceFrameDataCopyWithImpl<$Res>
           : importStatement // ignore: cast_nullable_to_non_nullable
               as String,
       dependencies: dependencies == freezed
-          ? _value.dependencies
+          ? _value._dependencies
           : dependencies // ignore: cast_nullable_to_non_nullable
               as List<String>,
     ));
@@ -144,7 +124,8 @@ class _$_WidgetbookDeviceFrameData implements _WidgetbookDeviceFrameData {
   _$_WidgetbookDeviceFrameData(
       {required this.name,
       required this.importStatement,
-      required this.dependencies});
+      required final List<String> dependencies})
+      : _dependencies = dependencies;
 
   factory _$_WidgetbookDeviceFrameData.fromJson(Map<String, dynamic> json) =>
       _$$_WidgetbookDeviceFrameDataFromJson(json);
@@ -153,8 +134,12 @@ class _$_WidgetbookDeviceFrameData implements _WidgetbookDeviceFrameData {
   final String name;
   @override
   final String importStatement;
+  final List<String> _dependencies;
   @override
-  final List<String> dependencies;
+  List<String> get dependencies {
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(_dependencies);
+  }
 
   @override
   String toString() {
@@ -165,27 +150,27 @@ class _$_WidgetbookDeviceFrameData implements _WidgetbookDeviceFrameData {
   bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _WidgetbookDeviceFrameData &&
+            other is _$_WidgetbookDeviceFrameData &&
             const DeepCollectionEquality().equals(other.name, name) &&
             const DeepCollectionEquality()
                 .equals(other.importStatement, importStatement) &&
             const DeepCollectionEquality()
-                .equals(other.dependencies, dependencies));
+                .equals(other._dependencies, _dependencies));
   }
 
+  @JsonKey(ignore: true)
   @override
   int get hashCode => Object.hash(
       runtimeType,
       const DeepCollectionEquality().hash(name),
       const DeepCollectionEquality().hash(importStatement),
-      const DeepCollectionEquality().hash(dependencies));
+      const DeepCollectionEquality().hash(_dependencies));
 
   @JsonKey(ignore: true)
   @override
-  _$WidgetbookDeviceFrameDataCopyWith<_WidgetbookDeviceFrameData>
-      get copyWith =>
-          __$WidgetbookDeviceFrameDataCopyWithImpl<_WidgetbookDeviceFrameData>(
-              this, _$identity);
+  _$$_WidgetbookDeviceFrameDataCopyWith<_$_WidgetbookDeviceFrameData>
+      get copyWith => __$$_WidgetbookDeviceFrameDataCopyWithImpl<
+          _$_WidgetbookDeviceFrameData>(this, _$identity);
 
   @override
   Map<String, dynamic> toJson() {
@@ -195,21 +180,21 @@ class _$_WidgetbookDeviceFrameData implements _WidgetbookDeviceFrameData {
 
 abstract class _WidgetbookDeviceFrameData implements WidgetbookDeviceFrameData {
   factory _WidgetbookDeviceFrameData(
-      {required String name,
-      required String importStatement,
-      required List<String> dependencies}) = _$_WidgetbookDeviceFrameData;
+      {required final String name,
+      required final String importStatement,
+      required final List<String> dependencies}) = _$_WidgetbookDeviceFrameData;
 
   factory _WidgetbookDeviceFrameData.fromJson(Map<String, dynamic> json) =
       _$_WidgetbookDeviceFrameData.fromJson;
 
   @override
-  String get name;
+  String get name => throw _privateConstructorUsedError;
   @override
-  String get importStatement;
+  String get importStatement => throw _privateConstructorUsedError;
   @override
-  List<String> get dependencies;
+  List<String> get dependencies => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  _$WidgetbookDeviceFrameDataCopyWith<_WidgetbookDeviceFrameData>
+  _$$_WidgetbookDeviceFrameDataCopyWith<_$_WidgetbookDeviceFrameData>
       get copyWith => throw _privateConstructorUsedError;
 }