+++ packages/widgetbook_generator/lib/models/widgetbook_use_case_builder_data.freezed.dart
@@ -1,5 +1,6 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
 part of 'widgetbook_use_case_builder_data.dart';
@@ -11,35 +12,12 @@ part of 'widgetbook_use_case_builder_data.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
 WidgetbookUseCaseBuilderData _$WidgetbookUseCaseBuilderDataFromJson(
     Map<String, dynamic> json) {
   return _WidgetbookUseCaseBuilderData.fromJson(json);
 }
-
-/// @nodoc
-class _$WidgetbookUseCaseBuilderDataTearOff {
-  const _$WidgetbookUseCaseBuilderDataTearOff();
-
-  _WidgetbookUseCaseBuilderData call(
-      {required String name,
-      required String importStatement,
-      required List<String> dependencies}) {
-    return _WidgetbookUseCaseBuilderData(
-      name: name,
-      importStatement: importStatement,
-      dependencies: dependencies,
-    );
-  }
-
-  WidgetbookUseCaseBuilderData fromJson(Map<String, Object?> json) {
-    return WidgetbookUseCaseBuilderData.fromJson(json);
-  }
-}
-
-/// @nodoc
-const $WidgetbookUseCaseBuilderData = _$WidgetbookUseCaseBuilderDataTearOff();
 
 /// @nodoc
 mixin _$WidgetbookUseCaseBuilderData {
@@ -95,28 +73,28 @@ class _$WidgetbookUseCaseBuilderDataCopyWithImpl<$Res>
 }
 
 /// @nodoc
-abstract class _$WidgetbookUseCaseBuilderDataCopyWith<$Res>
+abstract class _$$_WidgetbookUseCaseBuilderDataCopyWith<$Res>
     implements $WidgetbookUseCaseBuilderDataCopyWith<$Res> {
-  factory _$WidgetbookUseCaseBuilderDataCopyWith(
-          _WidgetbookUseCaseBuilderData value,
-          $Res Function(_WidgetbookUseCaseBuilderData) then) =
-      __$WidgetbookUseCaseBuilderDataCopyWithImpl<$Res>;
+  factory _$$_WidgetbookUseCaseBuilderDataCopyWith(
+          _$_WidgetbookUseCaseBuilderData value,
+          $Res Function(_$_WidgetbookUseCaseBuilderData) then) =
+      __$$_WidgetbookUseCaseBuilderDataCopyWithImpl<$Res>;
   @override
   $Res call({String name, String importStatement, List<String> dependencies});
 }
 
 /// @nodoc
-class __$WidgetbookUseCaseBuilderDataCopyWithImpl<$Res>
+class __$$_WidgetbookUseCaseBuilderDataCopyWithImpl<$Res>
     extends _$WidgetbookUseCaseBuilderDataCopyWithImpl<$Res>
-    implements _$WidgetbookUseCaseBuilderDataCopyWith<$Res> {
-  __$WidgetbookUseCaseBuilderDataCopyWithImpl(
-      _WidgetbookUseCaseBuilderData _value,
-      $Res Function(_WidgetbookUseCaseBuilderData) _then)
-      : super(_value, (v) => _then(v as _WidgetbookUseCaseBuilderData));
+    implements _$$_WidgetbookUseCaseBuilderDataCopyWith<$Res> {
+  __$$_WidgetbookUseCaseBuilderDataCopyWithImpl(
+      _$_WidgetbookUseCaseBuilderData _value,
+      $Res Function(_$_WidgetbookUseCaseBuilderData) _then)
+      : super(_value, (v) => _then(v as _$_WidgetbookUseCaseBuilderData));
 
   @override
-  _WidgetbookUseCaseBuilderData get _value =>
-      super._value as _WidgetbookUseCaseBuilderData;
+  _$_WidgetbookUseCaseBuilderData get _value =>
+      super._value as _$_WidgetbookUseCaseBuilderData;
 
   @override
   $Res call({
@@ -124,7 +102,7 @@ class __$WidgetbookUseCaseBuilderDataCopyWithImpl<$Res>
     Object? importStatement = freezed,
     Object? dependencies = freezed,
   }) {
-    return _then(_WidgetbookUseCaseBuilderData(
+    return _then(_$_WidgetbookUseCaseBuilderData(
       name: name == freezed
           ? _value.name
           : name // ignore: cast_nullable_to_non_nullable
@@ -134,7 +112,7 @@ class __$WidgetbookUseCaseBuilderDataCopyWithImpl<$Res>
           : importStatement // ignore: cast_nullable_to_non_nullable
               as String,
       dependencies: dependencies == freezed
-          ? _value.dependencies
+          ? _value._dependencies
           : dependencies // ignore: cast_nullable_to_non_nullable
               as List<String>,
     ));
@@ -147,7 +125,8 @@ class _$_WidgetbookUseCaseBuilderData implements _WidgetbookUseCaseBuilderData {
   _$_WidgetbookUseCaseBuilderData(
       {required this.name,
       required this.importStatement,
-      required this.dependencies});
+      required final List<String> dependencies})
+      : _dependencies = dependencies;
 
   factory _$_WidgetbookUseCaseBuilderData.fromJson(Map<String, dynamic> json) =>
       _$$_WidgetbookUseCaseBuilderDataFromJson(json);
@@ -156,8 +135,12 @@ class _$_WidgetbookUseCaseBuilderData implements _WidgetbookUseCaseBuilderData {
   final String name;
   @override
   final String importStatement;
+  final List<String> _dependencies;
   @override
-  final List<String> dependencies;
+  List<String> get dependencies {
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(_dependencies);
+  }
 
   @override
   String toString() {
@@ -168,26 +151,27 @@ class _$_WidgetbookUseCaseBuilderData implements _WidgetbookUseCaseBuilderData {
   bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _WidgetbookUseCaseBuilderData &&
+            other is _$_WidgetbookUseCaseBuilderData &&
             const DeepCollectionEquality().equals(other.name, name) &&
             const DeepCollectionEquality()
                 .equals(other.importStatement, importStatement) &&
             const DeepCollectionEquality()
-                .equals(other.dependencies, dependencies));
+                .equals(other._dependencies, _dependencies));
   }
 
+  @JsonKey(ignore: true)
   @override
   int get hashCode => Object.hash(
       runtimeType,
       const DeepCollectionEquality().hash(name),
       const DeepCollectionEquality().hash(importStatement),
-      const DeepCollectionEquality().hash(dependencies));
+      const DeepCollectionEquality().hash(_dependencies));
 
   @JsonKey(ignore: true)
   @override
-  _$WidgetbookUseCaseBuilderDataCopyWith<_WidgetbookUseCaseBuilderData>
-      get copyWith => __$WidgetbookUseCaseBuilderDataCopyWithImpl<
-          _WidgetbookUseCaseBuilderData>(this, _$identity);
+  _$$_WidgetbookUseCaseBuilderDataCopyWith<_$_WidgetbookUseCaseBuilderData>
+      get copyWith => __$$_WidgetbookUseCaseBuilderDataCopyWithImpl<
+          _$_WidgetbookUseCaseBuilderData>(this, _$identity);
 
   @override
   Map<String, dynamic> toJson() {
@@ -198,21 +182,22 @@ class _$_WidgetbookUseCaseBuilderData implements _WidgetbookUseCaseBuilderData {
 abstract class _WidgetbookUseCaseBuilderData
     implements WidgetbookUseCaseBuilderData {
   factory _WidgetbookUseCaseBuilderData(
-      {required String name,
-      required String importStatement,
-      required List<String> dependencies}) = _$_WidgetbookUseCaseBuilderData;
+          {required final String name,
+          required final String importStatement,
+          required final List<String> dependencies}) =
+      _$_WidgetbookUseCaseBuilderData;
 
   factory _WidgetbookUseCaseBuilderData.fromJson(Map<String, dynamic> json) =
       _$_WidgetbookUseCaseBuilderData.fromJson;
 
   @override
-  String get name;
+  String get name => throw _privateConstructorUsedError;
   @override
-  String get importStatement;
+  String get importStatement => throw _privateConstructorUsedError;
   @override
-  List<String> get dependencies;
+  List<String> get dependencies => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  _$WidgetbookUseCaseBuilderDataCopyWith<_WidgetbookUseCaseBuilderData>
+  _$$_WidgetbookUseCaseBuilderDataCopyWith<_$_WidgetbookUseCaseBuilderData>
       get copyWith => throw _privateConstructorUsedError;
 }