+++ packages/widgetbook_generator/lib/resolvers/app_builder_resolver.dart
@@ -1,0 +1,19 @@
+import 'package:widgetbook_annotation/widgetbook_annotation.dart';
+import 'package:widgetbook_generator/models/widgetbook_app_builder_data.dart';
+import 'package:widgetbook_generator/resolvers/builder_function_resolver.dart';
+
+class AppBuilderResolver extends BuilderFunctionResolver<WidgetbookAppBuilder> {
+  AppBuilderResolver()
+      : super(
+          createData: (
+            name,
+            imports,
+            dependencies,
+          ) =>
+              WidgetbookAppBuilderData(
+            name: name,
+            importStatement: imports,
+            dependencies: dependencies,
+          ),
+        );
+}